# static/swagger_oas.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IEQ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
...
  "paths": {
...
    "/order/get_all_orders": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "Get_all_orders",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoicmFtIiwiZW1haWwiOiJyYW1wQHRlc3QuY29tIiwiaWQiOiJjNDZkMTM4Yzk0MTY0NmFiYTc1YWFiMWIyYjYwNzZjYiIsInVzZXJfdHlwZSI6Mn0.KQC0fQl7Z9rQLDwbqzbQhm9FpNKRv8nB9uDG-kV5rWc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/order/create_order": {
      "post": {
        "tags": [
          "default"
        ],
        "summary": "Create order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "products": [
                    {
                      "id": 1,
                      "count": 2
                    },
                    {
                      "id": 2,
                      "count": 1
                    }
                  ],
                  "address": "1kjcdscjkvdsvwrew",
                  "phone": "9214821402"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoicmFtIiwiZW1haWwiOiJyYW1wQHRlc3QuY29tIiwiaWQiOiJjNDZkMTM4Yzk0MTY0NmFiYTc1YWFiMWIyYjYwNzZjYiIsInVzZXJfdHlwZSI6Mn0.KQC0fQl7Z9rQLDwbqzbQhm9FpNKRv8nB9uDG-kV5rWc"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "application/json"
          }
        ],
        "responses": {
          "201": {
            "description": "CREATED",
            "headers": {
              "Server": {
                "schema": {
                  "type": "number",
                  "example": "Werkzeug/3.0.1 Python/3.9.6"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Wed, 28 Feb 2024 06:32:12 GMT"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "1015"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "address": "1kjcdscjkvdsvwrew",
                  "deleted": false,
                  "order_id": "2a5b9d18cd3c480b94f085b0ccbe3936",
                  "order_status": 1,
                  "phone": "9214821402",
                  "products": [
                    {
                      "category": "Clothing",
                      "category_id": "61c56f61b7fd447496248ced88879515",
                      "count": 2,
                      "deleted": false,
                      "description": "This is a description for product 544. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
                      "id": 1,
                      "image": "https://picsum.photos/200/200?random=89",
                      "name": "Product 214",
                      "price": 1252.31,
                      "rating": 4.4,
                      "reviews": [],
                      "stock": 30,
                      "total": 2504.62
                    },
                    {
                      "category": "Electronics",
                      "category_id": "6034e1c2fa0d4c55a00615a4253f321d",
                      "count": 1,
                      "deleted": false,
                      "description": "This is a description for product 730. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
                      "id": 2,
                      "image": "https://picsum.photos/200/200?random=724",
                      "name": "Product 445",
                      "price": 3983.51,
                      "rating": 3.1,
                      "reviews": [],
                      "stock": 60,
                      "total": 3983.51
                    }
                  ],
                  "total": 6488.13
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "headers": {
              "Server": {
                "schema": {
                  "type": "number",
                  "example": "Werkzeug/3.0.1 Python/3.9.6"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Wed, 28 Feb 2024 06:33:03 GMT"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "33"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example-0": {
                    "summary": "Product required",
                    "value": {
                      "error": "products is required"
                    }
                  },
                  "example-1": {
                    "summary": "Invalie product ordered",
                    "value": {
                      "error": "Some invalid product ordered"
                    }
                  },
                  "example-2": {
                    "summary": "Address Reuired",
                    "value": {
                      "error": "address is required"
                    }
                  },
                  "example-3": {
                    "summary": "Phone required",
                    "value": {
                      "error": "phone is required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}